+++ core/src/br/unb/shooter/entity/Wall.java
@@ -2,7 +2,7 @@ package br.unb.shooter.entity;
 
 public class Wall extends Entity {
 
-    public Wall(Float x, Float y, Integer w, Integer h) {
+    public Wall(Float x, Float y, Float w, Float h) {
         this.positionX = x;
         this.positionY = y;
         this.width = w;
